% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpf.R
\name{bpf}
\alias{bpf}
\title{The Bootstrap Particle Filter}
\usage{
bpf(N, y, tk_list, alpha)
}
\arguments{
\item{N}{Number of particles}

\item{y}{Measurement process as T by N matrix}

\item{tk_list}{A length T list. The t-th element is a list containing the mean and covariance function of the Gaussian transition
kernel, and the potential function, at time t. See example.}

\item{alpha}{A number controls adaptive resampling.}
}
\value{
A list containing the particles (N by p matrix), weights (N by p matrix) and a vector indicating at which times
the algorithm has resampled.
}
\description{
The Bootstrap Particle Filter
}
\examples{
A<-diag(c(0.9,0.7,0.6,0.3,0))
A[lower.tri(A)]<-c(0.3,0.1,0.4,0.1,0.2,0.1,0.2,0.1,0.5,0.2)
LG_5_x_mean<-function(x){
 return(A\%*\%x)
}
LG_5_x_cov<-function(x){
return(diag(rep(1,5)))
}
LG_5_potential_log<-function(x,y){
 return(mvnfast::dmvn(y,x,diag(rep(0.25,5)),log = T))
}
tk_list_LG_5<-vector(mode = "list",101)
for (i in 1:101) {
 tk_list_LG_5[[i]]<-list(LG_5_x_mean,LG_5_x_cov,LG_5_potential_log)
}
T_data <- 100
data_obs<-LGSSM(A,T_data,0.25)
y<-data_obs$Y
bpf(500,y,tk_list_LG_5,0.5)
}
